Rm OpenAPI Generator Make target
This was used as an early POC to see if it would be a viable option as
the primary generator. It was quickly abandoned because the output was
quite messy. It seems to be better suited for path-heavy documents with
less complex schemas. As well, there's no Jsonnet generator which is the
primary language we'd like to generate code for. Creating a simpler
custom generator in Go was chosen over writing one for OpenAPI Generator
in Java.